Refactor move all activities into activities folder. Relink some activities.
