+++ app/src/main/AndroidManifest.xml
@@ -17,14 +17,14 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
         <activity android:name=".activities.PastProjectActivity"></activity>
-        <activity android:name=".DashboardPhase1Activity" />
-        <activity android:name=".DashboardPhase2Activity" />
-        <activity android:name=".DashboardPhase3Activity" />
-        <activity android:name=".DashboardPhase4Activity" />
-        <activity android:name=".DashboardPhase5Activity" />
-        <activity android:name=".DashboardPhase6Activity" />
-        <activity android:name=".DashboardPhase7Activity" />
-        <activity android:name=".DashboardPhase8Activity" />
+        <activity android:name=".activities.DashboardPhase1Activity" />
+        <activity android:name=".activities.DashboardPhase2Activity" />
+        <activity android:name=".activities.DashboardPhase3Activity" />
+        <activity android:name=".activities.DashboardPhase4Activity" />
+        <activity android:name=".activities.DashboardPhase5Activity" />
+        <activity android:name=".activities.DashboardPhase6Activity" />
+        <activity android:name=".activities.DashboardPhase7Activity" />
+        <activity android:name=".activities.DashboardPhase8Activity" />
         <activity android:name=".activities.DashBoardActivity" />
         <activity
             android:name=".activities.HomePageActivity"
