+++ app/src/main/java/com/bluesky/toa/activities/DashboardPhase2Activity.java
@@ -1,14 +1,11 @@
-package com.bluesky.toa;
+package com.bluesky.toa.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.bluesky.toa.activities.DashBoardActivity;
-import com.bluesky.toa.activities.FanProfileActivity;
-import com.bluesky.toa.activities.HomePageActivity;
-import com.bluesky.toa.activities.PastProjectActivity;
+import com.bluesky.toa.R;
 
 import androidx.appcompat.app.AppCompatActivity;
