+++ app/src/main/java/com/bluesky/toa/activities/DashboardPhase7Activity.java
@@ -1,25 +1,22 @@
-package com.bluesky.toa;
-
-import androidx.appcompat.app.AppCompatActivity;
+package com.bluesky.toa.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 
-import com.bluesky.toa.activities.DashBoardActivity;
-import com.bluesky.toa.activities.GuruProfileActivity;
-import com.bluesky.toa.activities.HomePageActivity;
-import com.bluesky.toa.activities.PastProjectActivity;
+import com.bluesky.toa.R;
 
-public class DashboardPhase1Activity extends AppCompatActivity {
+import androidx.appcompat.app.AppCompatActivity;
+
+public class DashboardPhase7Activity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_dashboard_phase1);
+        setContentView(R.layout.activity_dashboard_phase7);
 
-        Button projectDasboardBtn1 = findViewById(R.id.imageButton1_pdb_jj);
+        Button projectDasboardBtn1 = findViewById(R.id.imageButton7_pdb_jj);
 
         Button projectDasboardBtnEPF = findViewById(R.id.imageButton_edit_profile_pdb_jj);
         Button projectDasboardBtnHP = findViewById(R.id.imageButton_home_page_pdb_jj);
@@ -27,28 +24,28 @@ public class DashboardPhase1Activity extends AppCompatActivity {
 
         projectDasboardBtn1.setOnClickListener(new View.OnClickListener(){
             public void onClick(View v){
-                Intent intent = new Intent(DashboardPhase1Activity.this, DashBoardActivity.class);
+                Intent intent = new Intent(DashboardPhase7Activity.this, DashBoardActivity.class);
                 startActivity(intent);
             }
         });
 
         projectDasboardBtnEPF.setOnClickListener(new View.OnClickListener(){
             public void onClick(View v){
-                Intent intent = new Intent(DashboardPhase1Activity.this, GuruProfileActivity.class);
+                Intent intent = new Intent(DashboardPhase7Activity.this, GuruProfileActivity.class);
                 startActivity(intent);
             }
         });
 
         projectDasboardBtnHP.setOnClickListener(new View.OnClickListener(){
             public void onClick(View v){
-                Intent intent = new Intent(DashboardPhase1Activity.this, HomePageActivity.class);
+                Intent intent = new Intent(DashboardPhase7Activity.this, HomePageActivity.class);
                 startActivity(intent);
             }
         });
 
         projectDasboardBtnVPP.setOnClickListener(new View.OnClickListener(){
             public void onClick(View v){
-                Intent intent = new Intent(DashboardPhase1Activity.this, PastProjectActivity.class);
+                Intent intent = new Intent(DashboardPhase7Activity.this, PastProjectActivity.class);
                 startActivity(intent);
             }
         });
