+++ app/src/main/res/layout/activity_dashboard_phase3.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DashboardPhase3Activity">
+    tools:context=".activities.DashboardPhase3Activity">
 
     <TextView
         android:id="@+id/dash_board_p_title_jj"
